Start moving towards canonicalized database
