+++ generate_reflection/src/database.rs
@@ -4,12 +4,13 @@ use rbx_dom_weak::RbxValue;
 
 use crate::{
     api_dump::Dump,
-    canonical_properties::CanonicalPropertyDatabase,
+    reflection_types::RbxInstanceClass,
 };
 
 pub struct ReflectionDatabase {
     pub dump: Dump,
     pub default_properties: HashMap<String, HashMap<String, RbxValue>>,
-    pub canonical_properties: &'static CanonicalPropertyDatabase,
     pub studio_version: [u32; 4],
+
+    pub classes: HashMap<String, RbxInstanceClass>,
 }
